Fixed notification slide down animation
I noticed that when a notification is removed, surrounding notifications would jump abruptly due to the flex gap. I discovered that using a negative margin on individual elements can reverse the flex gap effect. This change prevents surrounding notifications from jumping when one is removed from the DOM, allowing for a smoother animation.